Actualizado a Prisma 1.8 y Añadido autorización

+++ src/application/post/post.mutation.js
@@ -1,12 +1,14 @@
+import { getUserId } from '../../utils/auth';
 import validate from '../../utils/validate';
 
 const Mutation = {};
 
-Mutation.createDraft = async (parent, { title, text, user }, ctx, info) => {
+Mutation.createDraft = async (parent, { title, text }, ctx, info) => {
+  const userID = getUserId(ctx);
+
   const fields = [
     { name: 'title', value: title, validate: ['required', 'noEmpty'] },
     { name: 'text', value: text, validate: ['required', 'noEmpty'] },
-    { name: 'user', value: user, validate: ['required', 'noEmpty'] },
   ];
 
   await validate(fields);
@@ -18,7 +20,7 @@ Mutation.createDraft = async (parent, { title, text, user }, ctx, info) => {
         text,
         isPublished: false,
         author: {
-          connect: { id: user },
+          connect: { id: userID },
         },
       },
     },
@@ -39,14 +41,30 @@ Mutation.publish = async (parent, { id }, ctx, info) => {
 
 
 Mutation.updatePost = async (parent, { id, title, text }, ctx, info) => {
+  // [1] `getUserId` Arroja un error si el usuario no está autenticado
+  // [2] Comprueba si el autor del post es el mismo que el usuario logueado
+  // [3] Solo si el usuario es el autor, podrá modificar el Post
+  // [4] Valido los datos
+  // [5] Actualizo el Post
+
+  const userId = getUserId(ctx); // [1]
+  const postExists = await ctx.db.exists.Post({ // [2]
+    id,
+    author: { id: userId },
+  });
+
+  if (!postExists) { // [3]
+    throw new Error(`Post not found or you're not the author`);
+  }
+
   const fields = [
     { name: 'title', value: title, validate: ['noEmpty'] },
     { name: 'text', value: text, validate: ['noEmpty'] },
   ];
 
-  await validate(fields);
+  await validate(fields); // [4]
 
-  return ctx.db.mutation.updatePost(
+  return ctx.db.mutation.updatePost( // [5]
     {
       where: { id },
       data: {
@@ -60,7 +78,22 @@ Mutation.updatePost = async (parent, { id, title, text }, ctx, info) => {
 
 
 Mutation.deletePost = async (parent, { id }, ctx, info) => {
-  return ctx.db.mutation.deletePost({ where: { id } }, info);
+  // [1] `getUserId` Arroja un error si el usuario no está autenticado
+  // [2] Comprueba si el autor del post es el mismo que el usuario logueado
+  // [3] Solo si el usuario es el autor, podrá modificar el Post
+  // [4] Actualizo el Post
+
+  const userId = getUserId(ctx); // [1]
+  const postExists = await ctx.db.exists.Post({ // [2]
+    id,
+    author: { id: userId },
+  });
+
+  if (!postExists) { // [3]
+    throw new Error(`Post not found or you're not the author`);
+  }
+
+  return ctx.db.mutation.deletePost({ where: { id } }, info); // [4]
 };
 
 
@@ -94,13 +127,15 @@ Mutation.deletePostAuthor = async (parent, { id, user }, ctx, info) => {
 };
 
 
-Mutation.likePost = async (parent, { id, user }, ctx, info) => {
+Mutation.likePost = async (parent, { id }, ctx, info) => {
+  const userID = getUserId(ctx);
+
   return ctx.db.mutation.updatePost(
     {
       where: { id },
       data: {
         likedBy: {
-          connect: { id: user },
+          connect: { id: userID },
         },
       },
     },
@@ -109,13 +144,15 @@ Mutation.likePost = async (parent, { id, user }, ctx, info) => {
 };
 
 
-Mutation.dislikePost = async (parent, { id, user }, ctx, info) => {
+Mutation.dislikePost = async (parent, { id }, ctx, info) => {
+  const userID = getUserId(ctx);
+
   return ctx.db.mutation.updatePost(
     {
       where: { id },
       data: {
         likedBy: {
-          delete: { id: user },
+          delete: { id: userID },
         },
       },
     },
